security/security-design/shared-assets/oci-security-health-check-forensics/requirements.txt
pandas
pyyaml
duckdb
oci
questionary
tqdm
gitpython
requests